backend/api/main.py
"""
AutoSec AI - Main FastAPI Application
Autonomous Cloud Security & Threat Mitigation Agent
"""

from fastapi import FastAPI, HTTPException
from fastapi.middleware.cors import CORSMiddleware
from pydantic import BaseModel
from datetime import datetime
from typing import Dict, List, Optional
import os
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# Initialize FastAPI app
app = FastAPI(
    title="AutoSec AI",
    description="Autonomous Cloud Security & Threat Mitigation Agent",
    version="0.1.0"
)

# Configure CORS
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],  
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# ============================================================================
# Pydantic Models (Data Schemas)
# ============================================================================

class HealthResponse(BaseModel):
    """Health check response model"""
    status: str
    timestamp: datetime
    version: str
    services: Dict[str, str]

class LogEvent(BaseModel):
    """Security log event model"""
    timestamp: datetime
    source_ip: str
    user_id: Optional[str] = None
    action: str
    resource: str
    status: str
    metadata: Optional[Dict] = {}

class ThreatAlert(BaseModel):
    """Threat alert model"""
    alert_id: str
    severity: str  # "low", "medium", "high"
    confidence: float  # 0.0 to 1.0
    threat_type: str
    description: str
    timestamp: datetime
    affected_resources: List[str]

### API Endpoints
@app.get("/")
async def root():
    """Root endpoint - API welcome message"""
    return {
        "message": "Welcome to AutoSec AI API",
        "version": "0.1.0",
        "docs": "/docs",
        "health": "/health"
    }

@app.get("/health", response_model=HealthResponse)
async def health_check():
    """
    Health check endpoint
    Returns the status of all system components
    """
    return HealthResponse(
        status="healthy",
        timestamp=datetime.now(),
        version="0.1.0",
        services={
            "api": "operational",
            "database": "pending",  
            "rag": "pending",       
            "agents": "pending"     
        }
    )

@app.get("/api/v1/status")
async def get_system_status():
    """
    Get current system status and statistics
    """
    return {
        "status": "running",
        "uptime": "0 hours",  # TODO: Implement actual uptime tracking
        "threats_detected_today": 0,
        "alerts_generated": 0,
        "last_scan": None
    }

@app.post("/api/v1/logs/ingest")
async def ingest_log(log: LogEvent):
    """
    Ingest a security log event
    This will be the entry point for log data
    """
    # TODO: Add actual log processing in Week 1 Day 4
    return {
        "status": "received",
        "log_id": f"log_{datetime.now().timestamp()}",
        "message": "Log ingested successfully",
        "log": log.dict()
    }

@app.get("/api/v1/threats")
async def get_threats(limit: int = 10):
    """
    Get recent threat alerts
    """
    # TODO: Connect to database to fetch real alerts
    return {
        "threats": [],
        "total": 0,
        "limit": limit,
        "message": "No threats detected yet - database connection pending"
    }

@app.get("/api/v1/threats/{alert_id}")
async def get_threat_detail(alert_id: str):
    """
    Get detailed information about a specific threat
    """
    # TODO: Implement threat retrieval from database
    raise HTTPException(
        status_code=404,
        detail=f"Threat {alert_id} not found - database not yet connected"
    )

@app.post("/api/v1/analyze")
async def analyze_log(log: LogEvent):
    """
    Analyze a log event for threats
    This will trigger the AI agent pipeline in Week 2
    """
    # TODO: Connect to AI agents for analysis
    return {
        "status": "analyzed",
        "threat_detected": False,
        "confidence": 0.0,
        "message": "AI agents not yet initialized - coming in Week 2"
    }

# ============================================================================
# Startup and Shutdown Events
# ============================================================================

@app.on_event("startup")
async def startup_event():
    """
    Run initialization tasks when the server starts
    """
    print("🚀 AutoSec AI starting up...")
    print(f"📝 API Documentation: http://localhost:8000/docs")
    print(f"🔍 Health Check: http://localhost:8000/health")
    
    # TODO: Initialize database connection
    # TODO: Load RAG vector store
    # TODO: Initialize AI agents

@app.on_event("shutdown")
async def shutdown_event():
    """
    Clean up resources when the server shuts down
    """
    print("👋 AutoSec AI shutting down...")
    # TODO: Close database connections
    # TODO: Save any pending data

# ============================================================================
# Run the server (for development)
# ============================================================================

if __name__ == "__main__":
    import uvicorn
    
    # Get configuration from environment variables
    host = os.getenv("API_HOST", "0.0.0.0")
    port = int(os.getenv("API_PORT", "8000"))
    reload = os.getenv("API_RELOAD", "true").lower() == "true"
    
    print(f"🛡️  Starting AutoSec AI on {host}:{port}")
    
    uvicorn.run(
        "main:app",
        host=host,
        port=port,
        reload=reload,  # Auto-reload on code changes (dev only)
        log_level="info"
    )