requirements.txt
# Core API Framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
python-multipart>=0.0.9

# LangChain & AI
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
openai>=1.0.0

# Vector Store & Embeddings
chromadb>=0.4.22
faiss-cpu>=1.7.4
sentence-transformers>=2.3.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0

# Database
sqlalchemy>=2.0.0

# Logging & Monitoring
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Data Visualization (for notebooks)
matplotlib>=3.7.0
seaborn>=0.12.0

# Jupyter (for exploration)
jupyter>=1.0.0
ipykernel>=6.25.0